houses live tracking state & remove all watch jobs
