+++ src/store/sagas/index.js
@@ -8,7 +8,7 @@ import { loginSaga, logoutSaga, loggedInSaga } from './authorizationSaga';
 import fetchCarLotsSaga from './fetchCarLotsSaga';
 import fetchHouseLotsSaga from './fetchHouseLotsSaga';
 import updateHouseWatchLotsSaga from './updateHouseWatchLotsSaga';
-import { watchHouseLotsSaga, checkWatchHouseLotsStateSaga } from './watchHouseLotsSaga';
+import { watchHouseLotsSaga, checkWatchHouseLotsStateSaga, updateHouseWatchFilterApplySaga } from './watchHouseLotsSaga';
 
 export default function* (service) {
   const rootSagas = [
@@ -24,6 +24,7 @@ export default function* (service) {
     updateHouseWatchLotsSaga(),
     watchHouseLotsSaga(),
     checkWatchHouseLotsStateSaga(),
+    updateHouseWatchFilterApplySaga(),
   ];
 
   yield all(rootSagas, service);
